+++ dune/common/iteratorrange.hh
@@ -14,10 +14,14 @@ namespace Dune {
    * existing containers that lack a standard begin(), end()
    * pair of member functions.
    *
+   * This supports to use a different type for the end iterator
+   * (a so called sentinel).
+   *
    * \tparam Iterator  The type of iterator
+   * \tparam Sentinel  The type of the end iterator (default=Iterator)
    * \ingroup CxxUtilities
    */
-  template<typename Iterator>
+  template<typename Iterator, typename Sentinel=Iterator>
   class IteratorRange
   {
 
@@ -27,13 +31,16 @@ namespace Dune {
     typedef Iterator iterator;
 
     //! The iterator belonging to this range.
+    typedef Sentinel sentinel;
+
+    //! The iterator belonging to this range.
     /**
      * This typedef is here mainly for compatibility reasons.
      */
     typedef Iterator const_iterator;
 
     //! Constructs an iterator range on [begin,end).
-    IteratorRange(const Iterator& begin, const Iterator& end)
+    IteratorRange(const Iterator& begin, const Sentinel& end)
       : _begin(begin)
       , _end(end)
     {}
@@ -49,7 +56,7 @@ namespace Dune {
     }
 
     //! Returns an iterator pointing past the end of the range.
-    iterator end() const
+    sentinel end() const
     {
       return _end;
     }
@@ -57,7 +64,7 @@ namespace Dune {
   private:
 
     Iterator _begin;
-    Iterator _end;
+    Sentinel _end;
 
   };